move more common utility functions to util.common
